+++ accounting/rbac/routing.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends, HTTPException
 from .endpoints import UserRoleCRUD
-from accounting.users.schemas import UserRead
+from accounting.schemas import UserRead, RoleRead
 userRole_router = APIRouter(
     prefix="/accounting/rbac",
     tags=["AAA->Accounting->RBAC"],
@@ -11,8 +11,15 @@ userRole_router = APIRouter(
 )
 
 userRole_router.add_api_route(
-    '/user/', 
+    '/user/roles/', 
     UserRoleCRUD.add_roles_to_user, 
     response_model=UserRead,
     summary='Add roles to user', 
+    methods=['put'])
+
+userRole_router.add_api_route(
+    '/role/users/', 
+    UserRoleCRUD.add_users_to_role, 
+    response_model=RoleRead,
+    summary='Add users to role', 
     methods=['put'])
