+++ accounting/roles/endpoints.py
@@ -1,6 +1,6 @@
 from loguru import logger
 from .models import Role
-from .schemas import (
+from accounting.schemas import (
     RoleCreate,
     RoleUpdate,
 )
